实现用户注册。 实现用户注册后自动订阅和首次推送。 部分 Id/Sid 改用 long 类型。

+++ Keylol.Models/AtRecord.cs
@@ -10,7 +10,7 @@ namespace Keylol.Models
 
         [Index(IsUnique = true, IsClustered = true)]
         [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
-        public int Sid { get; set; }
+        public long Sid { get; set; }
 
         public AtRecordEntryType EntryType { get; set; }
